pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "znote-mcp"
version = "1.3.0"
description = "A Zettelkasten knowledge management system as an MCP server"
authors = [
    {name = "Peter J. Herrel", email = "peterherrel@protonmail.com"},
    {name = "Alex Kechichian"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-frontmatter>=1.0.0",
    "markdown>=3.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
